backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai>=1.0.0
chromadb>=0.5.0
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
tavily-python>=0.7.0